+++ contracts/modules/interfaces/IexecPocoBoostAccessors.sol
@@ -18,7 +18,7 @@
 
 pragma solidity ^0.8.0;
 
-import "../../libs/IexecLibCore_v5.sol";
+import {IexecLibCore_v5} from "../../libs/IexecLibCore_v5.sol";
 
 interface IexecPocoBoostAccessors {
     function viewDealBoost(bytes32 id) external view returns (IexecLibCore_v5.DealBoost memory);