+++ CHANGELOG.md
@@ -72,7 +72,7 @@
 - Add dedicated Poco Boost accessors. (#206)
 - Refactor tests. (#177, #178, #179, #181, #188, #200, #209, #210, #211, #221)
 - Add mocha to use test explorer. (#180)
-- Reformat contracts (#222)
+- Reformat contracts (#222, #237)
 - Check balance in transfer operation (#230)
 
 ## v5.4.2